Convert the subjectAltName into an IP Address for comparison.
Convert the subjectAltName component to an IP Address and then
cast to a string prior to comparing to adding it to the set of
IP addresses found in the self-signed certificate properies.
Casting it directly to a str doesn't work as it is an OctetString
type and presents the data as the raw bytes of an IP address.

The ipaddress library already has the capability of parsing the
octet string bytes into an IP address and handles both IPv4 and
IPv6 forms.