Set statistics listener-address to 127.0.0.1.
If `listener-address` isn't set on the relation, the telegraf charm
defaults to using the unit's `private-address`; but
`monitoring_allowed_cidr` defaults to `127.0.0.1/32`, so if telegraf
tries to talk to haproxy on the unit's private address then it will get
a 403 error unless `monitoring_allowed_cidr` has been specifically
configured to allow the unit's private address.  Doing that is
inconvenient in bundles that are to be deployed in multiple network
environments.

The statistics relation has `scope: container`, so the other end of it
is always going to be on the same machine, and it makes sense to just
advertise 127.0.0.1.  This matches the behaviour of the content-cache
charm.